requirements.txt
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
numpy==1.26.4
pandas==2.2.2
pypdf==4.3.1
streamlit==1.37.1
transformers==4.44.2
accelerate==0.34.2
huggingface-hub==0.24.6
openai==1.48.0
uvicorn==0.30.6
fastapi==0.114.0
click==8.1.7
python-dotenv==1.0.1


